week2 movement and jump inplemented

+++ Assets/Scripts/Week_1/Week_1_Player.cs
@@ -8,7 +8,7 @@ public class Week_1_Player : MonoBehaviour
     private SpriteRenderer _spriteRenderer;
 
     private Vector3 Speed;
-    public int Velocity;
+    public float Velocity;
 
     RaycastHit hit;
     int RayDistance;
